rename shapes.hpp to shapes.h
